secured app

+++ server/index.js
@@ -17,6 +17,16 @@ app.use(express.json());
 
 // TODO - create middleware function isLoggedIn; use in the /api/auth/me route
 
+const isLoggedIn = async (req, res, next) => {
+  try {
+    req.user = await findUserByToken(req.headers.authorization);
+    next();
+  } catch (error) {
+    next();
+  }
+  
+}
+
 app.post("/api/auth/login", async (req, res, next) => {
   try {
     res.send(await authenticate(req.body));
@@ -25,9 +35,9 @@ app.post("/api/auth/login", async (req, res, next) => {
   }
 });
 
-app.get("/api/auth/me", async (req, res, next) => {
+app.get("/api/auth/me", isLoggedIn, async (req, res, next) => {
   try {
-    res.send(await findUserByToken(req.headers.authorization));
+    res.send(req.user);
   } catch (ex) {
     next(ex);
   }
@@ -50,10 +60,15 @@ app.get("/api/users", async (req, res, next) => {
 });
 
 // TODO - use isLoggedIn middleware
-app.get("/api/users/:id/userSkills", async (req, res, next) => {
+app.get("/api/users/:id/userSkills", isLoggedIn, async (req, res, next) => {
   try {
     // TODO - verify that the req.user.id on the route matches the user id in req.params
 
+    if(req.params.id != req.user.id) {
+      const error = Error('not authozired')
+      error.status = 401
+      throw error
+    }
     res.send(await fetchUserSkills(req.params.id));
   } catch (ex) {
     next(ex);
@@ -61,13 +76,13 @@ app.get("/api/users/:id/userSkills", async (req, res, next) => {
 });
 
 // TODO - use isLoggedIn middleware
-app.delete("/api/users/:userId/userSkills/:id", async (req, res, next) => {
+app.delete("/api/users/:userId/userSkills/:id", isLoggedIn, async (req, res, next) => {
   try {
-    // if (req.params.userId !== req.user.id) {
-    //   const error = Error("not authorized");
-    //   error.status = 401;
-    //   throw error;
-    // }
+    if(req.params.userId != req.user.id) {
+      const error = Error('not authozired')
+      error.status = 401
+      throw error
+    }
     await deleteUserSkill({ user_id: req.params.userId, id: req.params.id });
     res.sendStatus(204);
   } catch (ex) {
@@ -76,13 +91,13 @@ app.delete("/api/users/:userId/userSkills/:id", async (req, res, next) => {
 });
 
 // TODO - use isLoggedIn middleware
-app.post("/api/users/:id/userSkills", async (req, res, next) => {
+app.post("/api/users/:id/userSkills", isLoggedIn, async (req, res, next) => {
   try {
-    // if (req.params.id !== req.user.id) {
-    //   const error = Error("not authorized");
-    //   error.status = 401;
-    //   throw error;
-    // }
+    if (req.params.id !== req.user.id) {
+      const error = Error("not authorized");
+      error.status = 401;
+      throw error;
+    }
     res.status(201).send(
       await createUserSkill({
         user_id: req.params.id,
